backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
fsspec==2026.1.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pgvector==0.4.2
psycopg2-binary==2.9.11
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.0
SQLAlchemy==2.0.45
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.5
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
